Split identity server into v2 and v3
We make the IdentityServer object an abstraction layer between v2 and v3
APIs as the whole object. By doing this we don't have to check at the
beginning of every function whether to use v2 or v3. This will be
important when using auth plugins because the expected workflow is just
to issue the calls and see which ones fail with EndpointNotFound and so
we would have to reorder the code to possibly issue each call twice.

This changes the timing of discovery somewhat. Discovery is still
delayed but now it happens when the IdentityServer object is created
rather than when an IdentityServer object is accessed. In practice this
should have no impact as we simply care that it does not happen on
__init__ however it means that some tests need to have discovery entries
where they didn't previously.

Change-Id: Ie05d99946c2b1127992cd9a083a2068a6e80398b