Additional discovery changes
Move the discovery into the auth plugin itself. This is more inline with
how the rest of the plugins work in that if discovery doesn't tell you
that a URL is available for that version then we don't attempt to send
the query.

Because of this we are doing discovery checking before all our requests.
This response is cached so it is not an extra call. However this means
that we need to add proper discovery fixtures into the test file where
they might otherwise have been missing.

Change-Id: I19b2cd1aaa1e0b371f9e0c0b93147b70aeb61c47